refactor(functions): file-organization sweep over azure, functions and azure-mock
The resolved-notification shape is a model file under notification/, beside the resolver that returns it. azure and azure-mock report nothing: every export has a consumer, and nothing sits outside the subfolder that reads it.
