+++ apps/functions/src/models/notification/ResolvedNotification.ts
@@ -1,0 +1,7 @@
+export interface ResolvedNotification {
+  body: string;
+  icon?: null | string;
+  path: string;
+  title: string;
+  userIds: string[];
+}
